Implement a parser for Lisp data
This generates an AST from the tokenized input. The output of this will
be used for driving the interpreter. Representation still might need to
change a bit, but this seems like a somewhat sane starting point.

A lot of inspiration here comes from go's ast and parser packages. Once
I've got a basic interpreter up and running, I can revisit the
representation to either simplify or enhance the interpretation.