Reimplement parsing in terms of reading
Parse was useful for the initial proof, but this is less useful for
implementing evaluation.

A Lisp interpreter typically works in term of a loop that reads an
expression, evaluates it, and prints the result:

        (loop (print (eval (read))))

With this in mind, the renames Parse to Read and has it return an
expression as soon as one is completed.