Support for invoking primitive functions
This extends the environment with a number of primitive functions. The
idea here is to avoid special forms where not completely necessary.

Some benchmarks have been included to help track how well this
performs. I'm not sure what happens if this implementation is replaced
with judicious use of the reflect package, so I'm intrigued to see how
the numbers change. For the sake of interest, I'll contain with a naive
approach and play with different runtime representations
later (i.e. using the reflect package).