Primitive booleans and cons
In order to support expressions like (cons 1 nil), then it was necessary
to introduce T and NIL. This also necessitated NIL being made an atom.

This solves a couple of quirks that existed before, but also seems to
have introduced some overhead. Further investigation of this
representation is probably warranted.