Implement equal without reflection
We only have a few data types, so we can make a lot of assumptions. From
a benchmarking point of view, this is a pretty significant improvement.

I'm not sure what the implications of recursion are here, so need to
look into how the underlying Go runtime deals with it.