Implement the cond special form
This required a few enhancements to the treatment of symbols and
functions in order to correclty support equality. In particular, we
needed to ensure that `(quote t)` is equal to `t`. Without interning
symbols this was not the case.

Instead of implementing symbol internment, I've opted for improving the
logic for testing equality. I'm intruiged about the performance
implicitations of internment, so I'd like to use the current approach to
equality before make such drastic changes.

It did need some tweaks to functions to support equality as Go does not
natively support comparisons of functions. By wrapping them in
structures, we gain equality and a generalisation of invocation.