Move scanner into a separate package
It's getting to a point for me, where I'm losing sight of the runtime
portion of the code, so this is just an initial pass to start
extracting out the more concrete concepts into their own
packages.

The churn in this change suggests that the scanner has leaked into
more contexts than I expected, so there is probably some work to
provide some common entry points to reading Lisp data.