Simplify presentation of errors in scanner
This removes the special interface for configuring error reporting in
favour of introducing a new token type "Error". This simplifies the
interface since a client only needs to inspect tokens returned from
Next().

A side-effect is that the scanner has been simplified to reduce the
number of allocations overall, so this has had a very minor
improvement in performance.