Overhaul the representation of values
This started out as an effort to:

* Introduce error values.
* Move the reader into its own package.

As a result, the values are more self-contained, and conform to a
Value interface. The reader itself has a simpler interface since we
now simply return values in all cases. Sometimes they will be errors.

A lot less type switching is now required, and it's possible this is
reflected in the benchmark improvements. It's worth noting that the
Eval benchmark is useless until the interpreter is complete.

Pairs have also been removed for now, and lists are instead used for
representing lists. It's possible this will be a headache later on
when detecting improper lists, but it should be easier to refactor.