Implement preliminary special form handling
Overall the implementation is hairy, but the main goal is still to get
McCarthy's definition of Lisp self-hosting.

Environments have been improved to support extension so we can grow
the environment with the arguments passed to a function call.

One issue that has come up is the complexity introduced by
representing lists as arrays. Detecting improper lists is tricky, and
NIL doesn't play nicely (see case of `(lambda () 1)`). This needs
revisiting.