Allow definition of new symbols in the environment
This replaces extend in favour of the ability to define and update
symbols in the environment.

The advantage is we will be able to support the implementation of
label which typically needs to define an undefined symbol first before
updating the value with the resulting function.

As part of doing this, it seemed clearer to define the default
primitive operations in the value package. It is starting to smell
like the package structure probably isn't quite right.