Reintroduce cons cells as a list representation
The experiment with lists was a problem since it introduced ambiguity
for improper lists.

This still feels pretty klunky in terms of implemention, but thre
might be some better ways of destructuring the cells in Go, and still
give sensible error messages.

Also, some hairiness appears to still exist no matter the
implementation, since we really want `(eq NIL ())` to be true. This
might be solvable by defining nil differently.