Remove T and NIL from environment
These are core definitions, are used frequently, and we can avoid the
need to resolve these from the environment every time by checking for
them first when evaluating an environment.

I'm not certain if auto-quoting should stay. It's an extension on the
original defintion, but adds a _lot_ of convenience.