Make environment a first-class value
The intent is to provide some utilities for interacting with
environments so that we can easily call McCarthy's version of
EVAL. The ENVIRONMENT-BINDINGS function will be useful for cases like:

(eval
  ... expression ...
  (environment-bindings system-environment))

Trying to share logic between run.EvalString and run.Run provide to be
tricky since they have slightly different intended outcomes, so I've
decided to live with the duplication for now.

Also, in the process of testing the new environment capabilities, it's
starting to look like the auto-quoting of symbols is less useful. I've
left it alone for now, but it appears to be hiding a potential bug in
the evaluation of a Lambda. More investigation is probably required.