Add a Makefile as a test coverage harness
Typical tools haven't been particularly useful for coverage, but this
idea from the Nicolas Ruflin (see:
https://www.elastic.co/blog/code-coverage-for-your-golang-system-tests)
is pretty nifty for testing how much of the interpreter a specific
file actual interacts with.

The goal is to use this to sanity check the quality of the pure-Lisp
test suite that I'm working towards introducing.