Define APPLY for function application
This could potentially be simplified by exposing the underlying INVOKE
function and instead implementing this in the target
language. However, I still haven't figured out a good way to bake in
Lisp source to the interpreter.

For now, this is still targeting a pure Lisp test suite so that we can
simplify the Golang test suites.