Move evaluation to the run package
As I've tried to extend the core to expose things like the evaluator
and the reader, I'm finding that the value package feels like the
wrong place for evaluation.

The goal of this is to simplify the concerns of the `value` package to
only the representation of primitive types within the Lisp runtime.

Evaluation at this stage is largely a concern of the runtime, and
theoretically this will open up the possibility of configuring the
runtime to behave differently without changing the underlying types.

The system environment and primitives are yet to be moved, but this is
a starting point to shake out this approach.