value: Improve error reporting of APPLY
This was triggered by a linter warning:

> should replace loop with args = append(args, vs[1:len(vs)-1]...) (S1011)

Adding some tests made it apparent that the error reporting was a bit
unclear in more complex cases like "(APPLY FN A B)" where only the
last value expected to be a list was reported instead of all the
arguments.

The benchmark for the common case "(APPLY FN ARG-LIST)" is unimpacted,
but "(APPLY FN A B ARG-LIST)" has more allocations. Invocations and
lists probably warrant rethinking, but this smooths out usage in the
meantime.

It's also more apparent now that the package arrangement makes it
harder to get good test coverage by writing LISP directly. Another
thing to revisit soon so more can be written in the target
language. The test added here provides some more certainty in the
meantime.